requirements.txt
Django==3.0.2
psycopg2-binary==2.8.4
dj-database-url==0.5.0
django-cors-middleware==1.5.0
whitenoise==5.0.1
django-debug-toolbar==2.1
gunicorn==20.0.4
djangorestframework==3.11.0
djangorestframework-gis==0.15
djangorestframework-csv==2.1.0
django-filter==2.2.0
yapf==0.29.0
django-leaflet==0.26.0
python-dateutil==2.8.1
